correct SymbolProps
