+++ lib/src/symbols/props.tsx
@@ -1,4 +1,4 @@
-export interface SymbolProps extends React.HTMLAttributes<SVGElement> {
+export interface SymbolProps extends React.SVGProps<SVGSVGElement> {
   size?: number;
   color?: string;
 }
